+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to Rust's notion of
 
 ### Added
 - `impl subtle::ConstantTimeEq for note::Nullifier`
+- `impl Clone for note_encryption::CompactAction`
 
 ## [0.7.0] - 2024-01-26
 ### Licensing